crypto_bot/requirements.txt
requests
pandas
numpy
scikit-learn
tensorflow
yfinance
alpaca-py
ta-lib
matplotlib
python-dotenv
beautifulsoup4
